feat: refactor formula input
